+++ website/package.json
@@ -21,13 +21,7 @@
     "build": "yarn --cwd .. buildall && next build",
     "dev": "next dev",
     "export": "next build && next export",
-    "start": "next start",
-    "test": "mocha"
-  },
-  "mocha": {
-    "require": "tests/init.js",
-    "spec": "tests/**.ts",
-    "timeout": 9999999999
+    "start": "next start"
   },
   "dependencies": {
     "baby-i-am-faded": "4.0.14",
@@ -58,12 +52,10 @@
   "devDependencies": {
     "@gqlts/cli": "^3.2.10",
     "@types/js-cookie": "3.0.2",
-    "@types/mocha": "10.0.1",
     "@types/node": "18.11.18",
     "@types/prismjs": "1.26.0",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
-    "mocha": "10.2.0",
     "next-transpile-modules": "10.0.0",
     "sucrase": "3.29.0",
     "typescript": "4.9.4"
