--- website/tests/init.js
@@ -1,1 +1,0 @@
-require('sucrase/register')
